+++ server/cpho/views/indicator_metadata.py
@@ -162,7 +162,9 @@ class BenchmarkingForm(RequiredIfNotDeletedMixin, ModelForm):
 
             # year is required for all countries except OECD
             if year is None or year == "":
-                if oecd_country and not oecd_country.name_en in ["OECD"]:
+                if oecd_country and not oecd_country.name_en in [
+                    "OECD Average"
+                ]:
                     self.add_error(
                         "year",
                         tm("year_required"),